Refactor TimeseriesReturn and transformer
